@@ -1,6 +1,6 @@
 {
   "name": "rc-calendar",
-  "version": "9.6.0",
+  "version": "9.15.5",
   "description": "React Calendar",
   "keywords": [
     "react",
@@ -15,7 +15,8 @@
     "lib",
     "es",
     "assets/*.css",
-    "dist"
+    "dist",
+    "index.d.ts"
   ],
   "main": "lib/index",
   "module": "es/index",
@@ -46,6 +47,7 @@
     "pub": "rc-tools run pub --babel-runtime",
     "lint": "rc-tools run lint",
     "test": "jest",
+    "types": "tslint *.ts{,x} && tsc",
     "coverage": "jest --coverage"
   },
   "jest": {
@@ -60,11 +62,13 @@
     ],
     "transform": {
       "\\.jsx?$": "./node_modules/rc-tools/scripts/jestPreprocessor.js"
-    }
+    },
+    "testURL": "http://localhost"
   },
   "devDependencies": {
-    "async": "~2.5.0",
-    "bootstrap": "~3.3.2",
+    "@types/react": "^16.3.13",
+    "async": "~3.1.0",
+    "dtslint": "^0.9.0",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
     "enzyme-to-json": "^3.3.1",
@@ -72,12 +76,14 @@
     "mockdate": "^2.0.1",
     "pre-commit": "1.x",
     "rc-dialog": "^7.0.0",
-    "rc-select": "^7.1.0",
+    "rc-select": "^9.1.5",
     "rc-time-picker": "^3.1.0",
     "rc-tools": "^7.0.3",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
-    "react-test-renderer": "^16.0.0"
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2",
+    "react-test-renderer": "^16.0.0",
+    "tslint": "^5.9.1",
+    "typescript": "^3.4.1"
   },
   "pre-commit": [
     "lint"
@@ -85,10 +91,11 @@
   "dependencies": {
     "babel-runtime": "6.x",
     "classnames": "2.x",
-    "create-react-class": "^15.5.2",
     "moment": "2.x",
     "prop-types": "^15.5.8",
     "rc-trigger": "^2.2.0",
-    "rc-util": "^4.1.1"
-  }
+    "rc-util": "^4.1.1",
+    "react-lifecycles-compat": "^3.0.4"
+  },
+  "types": "index.d.ts"
 }
