@@ -2,17 +2,27 @@
 
 exports.__esModule = true;
 
+var _classCallCheck2 = require('babel-runtime/helpers/classCallCheck');
+
+var _classCallCheck3 = _interopRequireDefault(_classCallCheck2);
+
+var _possibleConstructorReturn2 = require('babel-runtime/helpers/possibleConstructorReturn');
+
+var _possibleConstructorReturn3 = _interopRequireDefault(_possibleConstructorReturn2);
+
+var _inherits2 = require('babel-runtime/helpers/inherits');
+
+var _inherits3 = _interopRequireDefault(_inherits2);
+
 var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
-var _createReactClass = require('create-react-class');
-
-var _createReactClass2 = _interopRequireDefault(_createReactClass);
-
 var _propTypes = require('prop-types');
 
 var _propTypes2 = _interopRequireDefault(_propTypes);
+
+var _reactLifecyclesCompat = require('react-lifecycles-compat');
 
 var _MonthTable = require('./MonthTable');
 
@@ -21,68 +31,67 @@ var _MonthTable2 = _interopRequireDefault(_MonthTable);
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
 
 function goYear(direction) {
-  var next = this.props.displayedValue.clone();
-  next.add(direction, 'year');
-  this.props.setDisplayedValue(next);
+  this.props.changeYear(direction);
 }
 
 function noop() {}
 
-var MonthPanel = (0, _createReactClass2['default'])({
-  displayName: 'MonthPanel',
+var MonthPanel = function (_React$Component) {
+  (0, _inherits3['default'])(MonthPanel, _React$Component);
 
-  propTypes: {
-    onChange: _propTypes2['default'].func,
-    disabledDate: _propTypes2['default'].func,
-    onSelect: _propTypes2['default'].func
-  },
+  function MonthPanel(props) {
+    (0, _classCallCheck3['default'])(this, MonthPanel);
 
-  getDefaultProps: function getDefaultProps() {
-    return {
-      onChange: noop,
-      onSelect: noop
+    var _this = (0, _possibleConstructorReturn3['default'])(this, _React$Component.call(this, props));
+
+    _this.setAndSelectValue = function (value) {
+      _this.setValue(value);
+      _this.props.onSelect(value);
     };
-  },
-  getInitialState: function getInitialState() {
-    var props = this.props;
-    // bind methods
-    this.nextYear = goYear.bind(this, 1);
-    this.previousYear = goYear.bind(this, -1);
-    this.prefixCls = props.rootPrefixCls + '-month-panel';
-    return {
+
+    _this.setValue = function (value) {
+      if ('value' in _this.props) {
+        _this.setState({
+          value: value
+        });
+      }
+    };
+
+    _this.nextYear = goYear.bind(_this, 1);
+    _this.previousYear = goYear.bind(_this, -1);
+    _this.prefixCls = props.rootPrefixCls + '-month-panel';
+
+    _this.state = {
       value: props.value || props.defaultValue
     };
-  },
-  componentWillReceiveProps: function componentWillReceiveProps(nextProps) {
-    if ('value' in nextProps) {
-      this.setState({
-        value: nextProps.value
-      });
+    return _this;
+  }
+
+  MonthPanel.getDerivedStateFromProps = function getDerivedStateFromProps(props) {
+    var newState = {};
+
+    if ('value' in props) {
+      newState = {
+        value: props.value
+      };
     }
-  },
-  setAndChangeValue: function setAndChangeValue(value) {
-    this.setValue(value);
-    this.props.onChange(value);
-  },
-  setAndSelectValue: function setAndSelectValue(value) {
-    this.setValue(value);
-    this.props.onSelect(value);
-  },
-  setValue: function setValue(value) {
-    if (!('value' in this.props)) {
-      this.setState({
-        value: value
-      });
-    }
-  },
-  render: function render() {
+
+    return newState;
+  };
+
+  MonthPanel.prototype.render = function render() {
     var props = this.props;
+    var locale = props.locale,
+        cellRender = props.cellRender,
+        contentRender = props.contentRender,
+        renderFooter = props.renderFooter;
+
     var displayedValue = props.displayedValue;
-    var cellRender = props.cellRender;
-    var contentRender = props.contentRender;
-    var locale = props.locale;
     var year = displayedValue.year();
     var prefixCls = this.prefixCls;
+
+    var footer = renderFooter && renderFooter('month');
+
     return _react2['default'].createElement(
       'div',
       { className: prefixCls, style: props.style },
@@ -136,11 +145,35 @@ var MonthPanel = (0, _createReactClass2['default'])({
             contentRender: contentRender,
             prefixCls: prefixCls
           })
+        ),
+        footer && _react2['default'].createElement(
+          'div',
+          { className: prefixCls + '-footer' },
+          footer
         )
       )
     );
-  }
-});
+  };
+
+  return MonthPanel;
+}(_react2['default'].Component);
+
+MonthPanel.propTypes = {
+  onChange: _propTypes2['default'].func,
+  disabledDate: _propTypes2['default'].func,
+  onSelect: _propTypes2['default'].func,
+  renderFooter: _propTypes2['default'].func,
+  rootPrefixCls: _propTypes2['default'].string,
+  value: _propTypes2['default'].object,
+  defaultValue: _propTypes2['default'].object
+};
+MonthPanel.defaultProps = {
+  onChange: noop,
+  onSelect: noop
+};
+
+
+(0, _reactLifecyclesCompat.polyfill)(MonthPanel);
 
 exports['default'] = MonthPanel;
 module.exports = exports['default'];
