@@ -2,13 +2,25 @@
 
 exports.__esModule = true;
 
+var _extends2 = require('babel-runtime/helpers/extends');
+
+var _extends3 = _interopRequireDefault(_extends2);
+
+var _classCallCheck2 = require('babel-runtime/helpers/classCallCheck');
+
+var _classCallCheck3 = _interopRequireDefault(_classCallCheck2);
+
+var _possibleConstructorReturn2 = require('babel-runtime/helpers/possibleConstructorReturn');
+
+var _possibleConstructorReturn3 = _interopRequireDefault(_possibleConstructorReturn2);
+
+var _inherits2 = require('babel-runtime/helpers/inherits');
+
+var _inherits3 = _interopRequireDefault(_inherits2);
+
 var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
-
-var _createReactClass = require('create-react-class');
-
-var _createReactClass2 = _interopRequireDefault(_createReactClass);
 
 var _propTypes = require('prop-types');
 
@@ -17,6 +29,8 @@ var _propTypes2 = _interopRequireDefault(_propTypes);
 var _KeyCode = require('rc-util/lib/KeyCode');
 
 var _KeyCode2 = _interopRequireDefault(_KeyCode);
+
+var _reactLifecyclesCompat = require('react-lifecycles-compat');
 
 var _CalendarHeader = require('./calendar/CalendarHeader');
 
@@ -28,79 +42,85 @@ var _CalendarFooter2 = _interopRequireDefault(_CalendarFooter);
 
 var _CalendarMixin = require('./mixin/CalendarMixin');
 
-var _CalendarMixin2 = _interopRequireDefault(_CalendarMixin);
-
 var _CommonMixin = require('./mixin/CommonMixin');
 
-var _CommonMixin2 = _interopRequireDefault(_CommonMixin);
+var _moment = require('moment');
+
+var _moment2 = _interopRequireDefault(_moment);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
 
-var MonthCalendar = (0, _createReactClass2['default'])({
-  displayName: 'MonthCalendar',
+var MonthCalendar = function (_React$Component) {
+  (0, _inherits3['default'])(MonthCalendar, _React$Component);
 
-  propTypes: {
-    monthCellRender: _propTypes2['default'].func,
-    dateCellRender: _propTypes2['default'].func
-  },
-  mixins: [_CommonMixin2['default'], _CalendarMixin2['default']],
+  function MonthCalendar(props) {
+    (0, _classCallCheck3['default'])(this, MonthCalendar);
 
-  getInitialState: function getInitialState() {
-    return { mode: 'month' };
-  },
-  onKeyDown: function onKeyDown(event) {
-    var keyCode = event.keyCode;
-    var ctrlKey = event.ctrlKey || event.metaKey;
-    var disabledDate = this.props.disabledDate;
-    var displayedValue = this.state.displayedValue;
+    var _this = (0, _possibleConstructorReturn3['default'])(this, _React$Component.call(this, props));
 
-    var value = displayedValue;
-    switch (keyCode) {
-      case _KeyCode2['default'].DOWN:
-        value = displayedValue.clone();
-        value.add(3, 'months');
-        break;
-      case _KeyCode2['default'].UP:
-        value = displayedValue.clone();
-        value.add(-3, 'months');
-        break;
-      case _KeyCode2['default'].LEFT:
-        value = displayedValue.clone();
-        if (ctrlKey) {
-          value.add(-1, 'years');
-        } else {
-          value.add(-1, 'months');
-        }
-        break;
-      case _KeyCode2['default'].RIGHT:
-        value = displayedValue.clone();
-        if (ctrlKey) {
-          value.add(1, 'years');
-        } else {
-          value.add(1, 'months');
-        }
-        break;
-      case _KeyCode2['default'].ENTER:
-        if (!disabledDate || !disabledDate(displayedValue)) {
-          this.onSelect(displayedValue);
-        }
+    _this.onKeyDown = function (event) {
+      var keyCode = event.keyCode;
+      var ctrlKey = event.ctrlKey || event.metaKey;
+      var disabledDate = _this.props.disabledDate;
+      var displayedValue = _this.state.displayedValue;
+
+      var value = displayedValue;
+      switch (keyCode) {
+        case _KeyCode2['default'].DOWN:
+          value = displayedValue.clone();
+          value.add(3, 'months');
+          break;
+        case _KeyCode2['default'].UP:
+          value = displayedValue.clone();
+          value.add(-3, 'months');
+          break;
+        case _KeyCode2['default'].LEFT:
+          value = displayedValue.clone();
+          if (ctrlKey) {
+            value.add(-1, 'years');
+          } else {
+            value.add(-1, 'months');
+          }
+          break;
+        case _KeyCode2['default'].RIGHT:
+          value = displayedValue.clone();
+          if (ctrlKey) {
+            value.add(1, 'years');
+          } else {
+            value.add(1, 'months');
+          }
+          break;
+        case _KeyCode2['default'].ENTER:
+          if (!disabledDate || !disabledDate(displayedValue)) {
+            _this.onSelect(displayedValue);
+          }
+          event.preventDefault();
+          return 1;
+        default:
+          return undefined;
+      }
+      if (value !== displayedValue) {
+        _this.setDisplayedValue(value);
         event.preventDefault();
         return 1;
-      default:
-        return undefined;
-    }
-    if (value !== displayedValue) {
-      this.setDisplayedValue(value);
-      event.preventDefault();
-      return 1;
-    }
-  },
-  handlePanelChange: function handlePanelChange(_, mode) {
-    if (mode !== 'date') {
-      this.setState({ mode: mode });
-    }
-  },
-  render: function render() {
+      }
+    };
+
+    _this.handlePanelChange = function (_, mode) {
+      if (mode !== 'date') {
+        _this.setState({ mode: mode });
+      }
+    };
+
+    _this.state = {
+      mode: 'month',
+      value: props.value || props.defaultValue || (0, _moment2['default'])(),
+      selectedValue: props.selectedValue || props.defaultSelectedValue
+    };
+    return _this;
+  }
+
+  MonthCalendar.prototype.render = function render() {
     var props = this.props,
         state = this.state;
     var mode = state.mode,
@@ -137,8 +157,19 @@ var MonthCalendar = (0, _createReactClass2['default'])({
       className: props.prefixCls + '-month-calendar',
       children: children
     });
-  }
-});
+  };
 
-exports['default'] = MonthCalendar;
+  return MonthCalendar;
+}(_react2['default'].Component);
+
+MonthCalendar.propTypes = (0, _extends3['default'])({}, _CalendarMixin.calendarMixinPropTypes, _CommonMixin.propType, {
+  monthCellRender: _propTypes2['default'].func,
+  value: _propTypes2['default'].object,
+  defaultValue: _propTypes2['default'].object,
+  selectedValue: _propTypes2['default'].object,
+  defaultSelectedValue: _propTypes2['default'].object,
+  disabledDate: _propTypes2['default'].func
+});
+MonthCalendar.defaultProps = (0, _extends3['default'])({}, _CommonMixin.defaultProp, _CalendarMixin.calendarMixinDefaultProps);
+exports['default'] = (0, _reactLifecyclesCompat.polyfill)((0, _CalendarMixin.calendarMixinWrapper)((0, _CommonMixin.commonMixinWrapper)(MonthCalendar)));
 module.exports = exports['default'];
