@@ -6,6 +6,18 @@ var _extends2 = require('babel-runtime/helpers/extends');
 
 var _extends3 = _interopRequireDefault(_extends2);
 
+var _classCallCheck2 = require('babel-runtime/helpers/classCallCheck');
+
+var _classCallCheck3 = _interopRequireDefault(_classCallCheck2);
+
+var _possibleConstructorReturn2 = require('babel-runtime/helpers/possibleConstructorReturn');
+
+var _possibleConstructorReturn3 = _interopRequireDefault(_possibleConstructorReturn2);
+
+var _inherits2 = require('babel-runtime/helpers/inherits');
+
+var _inherits3 = _interopRequireDefault(_inherits2);
+
 var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
@@ -13,10 +25,6 @@ var _react2 = _interopRequireDefault(_react);
 var _reactDom = require('react-dom');
 
 var _reactDom2 = _interopRequireDefault(_reactDom);
-
-var _createReactClass = require('create-react-class');
-
-var _createReactClass2 = _interopRequireDefault(_createReactClass);
 
 var _propTypes = require('prop-types');
 
@@ -44,40 +52,33 @@ var _TimePickerButton2 = _interopRequireDefault(_TimePickerButton);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
 
-var CalendarFooter = (0, _createReactClass2['default'])({
-  displayName: 'CalendarFooter',
+var CalendarFooter = function (_React$Component) {
+  (0, _inherits3['default'])(CalendarFooter, _React$Component);
 
-  propTypes: {
-    prefixCls: _propTypes2['default'].string,
-    showDateInput: _propTypes2['default'].bool,
-    disabledTime: _propTypes2['default'].any,
-    timePicker: _propTypes2['default'].element,
-    selectedValue: _propTypes2['default'].oneOfType([_propTypes2['default'].object, _propTypes2['default'].arrayOf(_propTypes2['default'].object)]),
-    showOk: _propTypes2['default'].bool,
-    onSelect: _propTypes2['default'].func,
-    value: _propTypes2['default'].oneOfType([_propTypes2['default'].object, _propTypes2['default'].arrayOf(_propTypes2['default'].object)]),
-    displayedValue: _propTypes2['default'].object,
-    renderFooter: _propTypes2['default'].func,
-    defaultValue: _propTypes2['default'].oneOfType([_propTypes2['default'].object, _propTypes2['default'].arrayOf(_propTypes2['default'].object)]),
-    multiple: _propTypes2['default'].bool
-  },
+  function CalendarFooter() {
+    (0, _classCallCheck3['default'])(this, CalendarFooter);
+    return (0, _possibleConstructorReturn3['default'])(this, _React$Component.apply(this, arguments));
+  }
 
-  onSelect: function onSelect(value) {
+  CalendarFooter.prototype.onSelect = function onSelect(value) {
     this.props.onSelect(value);
-  },
-  getRootDOMNode: function getRootDOMNode() {
+  };
+
+  CalendarFooter.prototype.getRootDOMNode = function getRootDOMNode() {
     return _reactDom2['default'].findDOMNode(this);
-  },
-  render: function render() {
+  };
+
+  CalendarFooter.prototype.render = function render() {
     var props = this.props;
     var value = props.value,
         prefixCls = props.prefixCls,
         showOk = props.showOk,
         timePicker = props.timePicker,
-        renderFooter = props.renderFooter;
+        renderFooter = props.renderFooter,
+        mode = props.mode;
 
     var footerEl = null;
-    var extraFooter = renderFooter();
+    var extraFooter = renderFooter && renderFooter(mode);
     if (props.showToday || timePicker || extraFooter) {
       var _cx;
 
@@ -95,24 +96,41 @@ var CalendarFooter = (0, _createReactClass2['default'])({
       }
 
       var footerBtn = void 0;
-      if (nowEl || timePickerBtn || okBtn) {
+      if (nowEl || timePickerBtn || okBtn || extraFooter) {
         footerBtn = _react2['default'].createElement(
           'span',
           { className: prefixCls + '-footer-btn' },
+          extraFooter,
           (0, _mapSelf2['default'])([nowEl, timePickerBtn, okBtn])
         );
       }
-      var cls = (0, _classnames2['default'])((_cx = {}, _cx[prefixCls + '-footer'] = true, _cx[prefixCls + '-footer-show-ok'] = okBtn, _cx));
+      var cls = (0, _classnames2['default'])(prefixCls + '-footer', (_cx = {}, _cx[prefixCls + '-footer-show-ok'] = okBtn, _cx));
       footerEl = _react2['default'].createElement(
         'div',
         { className: cls },
-        extraFooter,
         footerBtn
       );
     }
     return footerEl;
-  }
-});
+  };
 
+  return CalendarFooter;
+}(_react2['default'].Component);
+
+CalendarFooter.propTypes = {
+  prefixCls: _propTypes2['default'].string,
+  showDateInput: _propTypes2['default'].bool,
+  disabledTime: _propTypes2['default'].any,
+  timePicker: _propTypes2['default'].element,
+  selectedValue: _propTypes2['default'].oneOfType([_propTypes2['default'].object, _propTypes2['default'].arrayOf(_propTypes2['default'].object)]),
+  showOk: _propTypes2['default'].bool,
+  onSelect: _propTypes2['default'].func,
+  value: _propTypes2['default'].oneOfType([_propTypes2['default'].object, _propTypes2['default'].arrayOf(_propTypes2['default'].object)]),
+  displayedValue: _propTypes2['default'].object,
+  renderFooter: _propTypes2['default'].func,
+  mode: _propTypes2['default'].string,
+  defaultValue: _propTypes2['default'].oneOfType([_propTypes2['default'].object, _propTypes2['default'].arrayOf(_propTypes2['default'].object)]),
+  multiple: _propTypes2['default'].bool
+};
 exports['default'] = CalendarFooter;
 module.exports = exports['default'];
