@@ -2,13 +2,21 @@
 
 exports.__esModule = true;
 
+var _classCallCheck2 = require('babel-runtime/helpers/classCallCheck');
+
+var _classCallCheck3 = _interopRequireDefault(_classCallCheck2);
+
+var _possibleConstructorReturn2 = require('babel-runtime/helpers/possibleConstructorReturn');
+
+var _possibleConstructorReturn3 = _interopRequireDefault(_possibleConstructorReturn2);
+
+var _inherits2 = require('babel-runtime/helpers/inherits');
+
+var _inherits3 = _interopRequireDefault(_inherits2);
+
 var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
-
-var _createReactClass = require('create-react-class');
-
-var _createReactClass2 = _interopRequireDefault(_createReactClass);
 
 var _propTypes = require('prop-types');
 
@@ -48,149 +56,26 @@ function showIf(condition, el) {
   return condition ? el : null;
 }
 
-var CalendarHeader = (0, _createReactClass2['default'])({
-  displayName: 'CalendarHeader',
+var CalendarHeader = function (_React$Component) {
+  (0, _inherits3['default'])(CalendarHeader, _React$Component);
 
-  propTypes: {
-    prefixCls: _propTypes2['default'].string,
-    value: _propTypes2['default'].oneOfType([_propTypes2['default'].object, _propTypes2['default'].arrayOf(_propTypes2['default'].object)]),
-    displayedValue: _propTypes2['default'].object,
-    onValueChange: _propTypes2['default'].func,
-    showTimePicker: _propTypes2['default'].bool,
-    onPanelChange: _propTypes2['default'].func,
-    locale: _propTypes2['default'].object,
-    enablePrev: _propTypes2['default'].any,
-    enableNext: _propTypes2['default'].any,
-    disabledMonth: _propTypes2['default'].func,
-    disablePreviousMonth: _propTypes2['default'].bool,
-    disablePreviousYear: _propTypes2['default'].bool,
-    prevMonthIcon: _propTypes2['default'].node,
-    nextMonthIcon: _propTypes2['default'].node,
-    prevYearIcon: _propTypes2['default'].node,
-    nextYearIcon: _propTypes2['default'].node,
-    multiple: _propTypes2['default'].bool
-  },
+  function CalendarHeader(props) {
+    (0, _classCallCheck3['default'])(this, CalendarHeader);
 
-  getDefaultProps: function getDefaultProps() {
-    return {
-      enableNext: 1,
-      enablePrev: 1,
-      onPanelChange: function onPanelChange() {},
-      onValueChange: function onValueChange() {}
-    };
-  },
-  getInitialState: function getInitialState() {
-    this.nextMonth = goMonth.bind(this, 1);
-    this.previousMonth = goMonth.bind(this, -1);
-    this.nextYear = goYear.bind(this, 1);
-    this.previousYear = goYear.bind(this, -1);
-    return { yearPanelReferer: null };
-  },
-  onMonthSelect: function onMonthSelect(value) {
-    this.props.onPanelChange(value, 'date');
-    if (this.props.onMonthSelect) {
-      this.props.onMonthSelect(value);
-    } else {
-      this.props.setDisplayedValue(value);
-    }
-  },
-  onYearSelect: function onYearSelect(value) {
-    var referer = this.state.yearPanelReferer;
-    this.setState({ yearPanelReferer: null });
-    this.props.onPanelChange(value, referer);
-    this.props.setDisplayedValue(value);
-  },
-  onDecadeSelect: function onDecadeSelect(value) {
-    this.props.onPanelChange(value, 'year');
-    this.props.setDisplayedValue(value);
-  },
-  monthYearElement: function monthYearElement(showTimePicker) {
-    var _this = this;
+    var _this = (0, _possibleConstructorReturn3['default'])(this, _React$Component.call(this, props));
 
-    var props = this.props;
-    var prefixCls = props.prefixCls;
-    var locale = props.locale;
-    var displayedValue = props.displayedValue;
-    var localeData = displayedValue.localeData();
-    var monthBeforeYear = locale.monthBeforeYear;
-    var selectClassName = prefixCls + '-' + (monthBeforeYear ? 'my-select' : 'ym-select');
+    _initialiseProps.call(_this);
 
-    if (props.onMonthSelect) {
-      return _react2['default'].createElement(
-        'a',
-        {
-          className: prefixCls + '-month-select',
-          onClick: function onClick() {
-            props.onMonthSelect(displayedValue);
-          },
-          onMouseEnter: function onMouseEnter() {
-            props.onMonthMouseEnter(displayedValue);
-          },
-          onMouseLeave: props.onMonthMouseLeave
-        },
-        localeData.months(displayedValue),
-        ' ',
-        displayedValue.format(locale.yearFormat)
-      );
-    }
+    _this.nextMonth = goMonth.bind(_this, 1);
+    _this.previousMonth = goMonth.bind(_this, -1);
+    _this.nextYear = goYear.bind(_this, 1);
+    _this.previousYear = goYear.bind(_this, -1);
 
-    var year = _react2['default'].createElement(
-      'a',
-      {
-        className: prefixCls + '-year-select',
-        role: 'button',
-        onClick: showTimePicker ? null : function () {
-          return _this.showYearPanel('date');
-        },
-        title: locale.yearSelect
-      },
-      displayedValue.format(locale.yearFormat)
-    );
-    var month = _react2['default'].createElement(
-      'a',
-      {
-        className: prefixCls + '-month-select',
-        role: 'button',
-        onClick: showTimePicker ? null : this.showMonthPanel,
-        title: locale.monthSelect
-      },
-      locale.monthFormat ? displayedValue.format(locale.monthFormat) : localeData.monthsShort(displayedValue)
-    );
-    var day = void 0;
-    if (showTimePicker) {
-      day = _react2['default'].createElement(
-        'a',
-        {
-          className: prefixCls + '-day-select',
-          role: 'button'
-        },
-        displayedValue.format(locale.dayFormat)
-      );
-    }
-    var my = [];
-    if (monthBeforeYear) {
-      my = [month, day, year];
-    } else {
-      my = [year, month, day];
-    }
-    return _react2['default'].createElement(
-      'span',
-      { className: selectClassName },
-      (0, _mapSelf2['default'])(my)
-    );
-  },
-  showMonthPanel: function showMonthPanel() {
-    // null means that users' interaction doesn't change value
-    this.props.onPanelChange(null, 'month');
-  },
-  showYearPanel: function showYearPanel(referer) {
-    this.setState({ yearPanelReferer: referer });
-    this.props.onPanelChange(null, 'year');
-  },
-  showDecadePanel: function showDecadePanel() {
-    this.props.onPanelChange(null, 'decade');
-  },
-  render: function render() {
+    _this.state = { yearPanelReferer: null };
+    return _this;
+  }
+
+  CalendarHeader.prototype.render = function render() {
     var _this2 = this;
 
     var props = this.props;
@@ -203,6 +88,7 @@ var CalendarHeader = (0, _createReactClass2['default'])({
         enableNext = props.enableNext,
         enablePrev = props.enablePrev,
         disabledMonth = props.disabledMonth,
+        renderFooter = props.renderFooter,
         disablePreviousMonth = props.disablePreviousMonth,
         disablePreviousYear = props.disablePreviousYear,
         prevMonthIcon = props.prevMonthIcon,
@@ -215,6 +101,7 @@ var CalendarHeader = (0, _createReactClass2['default'])({
     if (mode === 'month') {
       panel = _react2['default'].createElement(_MonthPanel2['default'], {
         locale: locale,
+        value: value,
         defaultValue: value,
         displayedValue: displayedValue,
         rootPrefixCls: prefixCls,
@@ -225,6 +112,8 @@ var CalendarHeader = (0, _createReactClass2['default'])({
         disabledDate: disabledMonth,
         cellRender: props.monthCellRender,
         contentRender: props.monthCellContentRender,
+        renderFooter: renderFooter,
+        changeYear: this.changeYear,
         setDisplayedValue: props.setDisplayedValue
       });
     }
@@ -236,6 +125,7 @@ var CalendarHeader = (0, _createReactClass2['default'])({
         rootPrefixCls: prefixCls,
         onSelect: this.onYearSelect,
         onDecadePanelShow: this.showDecadePanel,
+        renderFooter: renderFooter,
         setDisplayedValue: props.setDisplayedValue
       });
     }
@@ -246,6 +136,7 @@ var CalendarHeader = (0, _createReactClass2['default'])({
         displayedValue: displayedValue,
         rootPrefixCls: prefixCls,
         onSelect: this.onDecadeSelect,
+        renderFooter: renderFooter,
         setDisplayedValue: props.setDisplayedValue
       });
     }
@@ -306,8 +197,161 @@ var CalendarHeader = (0, _createReactClass2['default'])({
       ),
       panel
     );
-  }
-});
+  };
+
+  return CalendarHeader;
+}(_react2['default'].Component);
+
+CalendarHeader.propTypes = {
+  prefixCls: _propTypes2['default'].string,
+  value: _propTypes2['default'].oneOfType([_propTypes2['default'].object, _propTypes2['default'].arrayOf(_propTypes2['default'].object)]),
+  displayedValue: _propTypes2['default'].object,
+  onValueChange: _propTypes2['default'].func,
+  showTimePicker: _propTypes2['default'].bool,
+  onPanelChange: _propTypes2['default'].func,
+  locale: _propTypes2['default'].object,
+  enablePrev: _propTypes2['default'].any,
+  enableNext: _propTypes2['default'].any,
+  disabledMonth: _propTypes2['default'].func,
+  renderFooter: _propTypes2['default'].func,
+  onMonthSelect: _propTypes2['default'].oneOfType([_propTypes2['default'].func, _propTypes2['default'].bool]),
+  disablePreviousMonth: _propTypes2['default'].bool,
+  disablePreviousYear: _propTypes2['default'].bool,
+  prevMonthIcon: _propTypes2['default'].node,
+  nextMonthIcon: _propTypes2['default'].node,
+  prevYearIcon: _propTypes2['default'].node,
+  nextYearIcon: _propTypes2['default'].node,
+  multiple: _propTypes2['default'].bool,
+  setDisplayedValue: _propTypes2['default'].func
+};
+CalendarHeader.defaultProps = {
+  enableNext: 1,
+  enablePrev: 1,
+  onPanelChange: function onPanelChange() {},
+  onValueChange: function onValueChange() {}
+};
+
+var _initialiseProps = function _initialiseProps() {
+  var _this3 = this;
+
+  this.onMonthSelect = function (value) {
+    _this3.props.onPanelChange(value, 'date');
+    if (_this3.props.onMonthSelect) {
+      _this3.props.onMonthSelect(value);
+    } else {
+      _this3.props.setDisplayedValue(value);
+    }
+  };
+
+  this.onYearSelect = function (value) {
+    var referer = _this3.state.yearPanelReferer;
+    _this3.setState({ yearPanelReferer: null });
+    _this3.props.onPanelChange(value, referer);
+    _this3.props.setDisplayedValue(value);
+  };
+
+  this.onDecadeSelect = function (value) {
+    _this3.props.onPanelChange(value, 'year');
+    _this3.props.setDisplayedValue(value);
+  };
+
+  this.changeYear = function (direction) {
+    if (direction > 0) {
+      _this3.nextYear();
+    } else {
+      _this3.previousYear();
+    }
+  };
+
+  this.monthYearElement = function (showTimePicker) {
+    var props = _this3.props;
+    var prefixCls = props.prefixCls;
+    var locale = props.locale;
+    var displayedValue = props.displayedValue;
+    var localeData = displayedValue.localeData();
+    var monthBeforeYear = locale.monthBeforeYear;
+    var selectClassName = prefixCls + '-' + (monthBeforeYear ? 'my-select' : 'ym-select');
+    var timeClassName = showTimePicker ? ' ' + prefixCls + '-time-status' : '';
+
+    if (props.onMonthSelect) {
+      return _react2['default'].createElement(
+        'a',
+        {
+          className: prefixCls + '-month-select',
+          onClick: function onClick() {
+            props.onMonthSelect(displayedValue);
+          },
+          onMouseEnter: function onMouseEnter() {
+            props.onMonthMouseEnter(displayedValue);
+          },
+          onMouseLeave: props.onMonthMouseLeave
+        },
+        localeData.months(displayedValue),
+        ' ',
+        displayedValue.format(locale.yearFormat)
+      );
+    }
+
+    var year = _react2['default'].createElement(
+      'a',
+      {
+        className: prefixCls + '-year-select' + timeClassName,
+        role: 'button',
+        onClick: showTimePicker ? null : function () {
+          return _this3.showYearPanel('date');
+        },
+        title: showTimePicker ? null : locale.yearSelect
+      },
+      displayedValue.format(locale.yearFormat)
+    );
+    var month = _react2['default'].createElement(
+      'a',
+      {
+        className: prefixCls + '-month-select' + timeClassName,
+        role: 'button',
+        onClick: showTimePicker ? null : _this3.showMonthPanel,
+        title: showTimePicker ? null : locale.monthSelect
+      },
+      locale.monthFormat ? displayedValue.format(locale.monthFormat) : localeData.monthsShort(displayedValue)
+    );
+    var day = void 0;
+    if (showTimePicker) {
+      day = _react2['default'].createElement(
+        'a',
+        {
+          className: prefixCls + '-day-select' + timeClassName,
+          role: 'button'
+        },
+        displayedValue.format(locale.dayFormat)
+      );
+    }
+    var my = [];
+    if (monthBeforeYear) {
+      my = [month, day, year];
+    } else {
+      my = [year, month, day];
+    }
+    return _react2['default'].createElement(
+      'span',
+      { className: selectClassName },
+      (0, _mapSelf2['default'])(my)
+    );
+  };
+
+  this.showMonthPanel = function () {
+    // null means that users' interaction doesn't change value
+    _this3.props.onPanelChange(null, 'month');
+  };
+
+  this.showYearPanel = function (referer) {
+    _this3.setState({ yearPanelReferer: referer });
+    _this3.props.onPanelChange(null, 'year');
+  };
+
+  this.showDecadePanel = function () {
+    _this3.props.onPanelChange(null, 'decade');
+  };
+};
 
 exports['default'] = CalendarHeader;
 module.exports = exports['default'];
