@@ -2,13 +2,21 @@
 
 exports.__esModule = true;
 
+var _classCallCheck2 = require('babel-runtime/helpers/classCallCheck');
+
+var _classCallCheck3 = _interopRequireDefault(_classCallCheck2);
+
+var _possibleConstructorReturn2 = require('babel-runtime/helpers/possibleConstructorReturn');
+
+var _possibleConstructorReturn3 = _interopRequireDefault(_possibleConstructorReturn2);
+
+var _inherits2 = require('babel-runtime/helpers/inherits');
+
+var _inherits3 = _interopRequireDefault(_inherits2);
+
 var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
-
-var _createReactClass = require('create-react-class');
-
-var _createReactClass2 = _interopRequireDefault(_createReactClass);
 
 var _propTypes = require('prop-types');
 
@@ -48,27 +56,15 @@ function getIdFromDate(date) {
   return 'rc-calendar-' + date.year() + '-' + date.month() + '-' + date.date();
 }
 
-var DateTBody = (0, _createReactClass2['default'])({
-  displayName: 'DateTBody',
+var DateTBody = function (_React$Component) {
+  (0, _inherits3['default'])(DateTBody, _React$Component);
 
-  propTypes: {
-    contentRender: _propTypes2['default'].func,
-    dateRender: _propTypes2['default'].func,
-    disabledDate: _propTypes2['default'].func,
-    prefixCls: _propTypes2['default'].string,
-    selectedValue: _propTypes2['default'].oneOfType([_propTypes2['default'].object, _propTypes2['default'].arrayOf(_propTypes2['default'].object)]),
-    value: _propTypes2['default'].oneOfType([_propTypes2['default'].object, _propTypes2['default'].arrayOf(_propTypes2['default'].object)]),
-    hoverValue: _propTypes2['default'].any,
-    showWeekNumber: _propTypes2['default'].bool,
-    multiple: _propTypes2['default'].bool
-  },
+  function DateTBody() {
+    (0, _classCallCheck3['default'])(this, DateTBody);
+    return (0, _possibleConstructorReturn3['default'])(this, _React$Component.apply(this, arguments));
+  }
 
-  getDefaultProps: function getDefaultProps() {
-    return {
-      hoverValue: []
-    };
-  },
-  render: function render() {
+  DateTBody.prototype.render = function render() {
     var props = this.props;
     var contentRender = props.contentRender,
         prefixCls = props.prefixCls,
@@ -112,6 +108,7 @@ var DateTBody = (0, _createReactClass2['default'])({
     var lastMonth1 = month1.clone();
     lastMonth1.add(0 - lastMonthDiffDay, 'days');
     var passed = 0;
+
     for (iIndex = 0; iIndex < _DateConstants2['default'].DATE_ROW_COUNT; iIndex++) {
       for (jIndex = 0; jIndex < _DateConstants2['default'].DATE_COL_COUNT; jIndex++) {
         current = lastMonth1;
@@ -201,11 +198,15 @@ var DateTBody = (0, _createReactClass2['default'])({
                 cls += ' ' + selectedStartDateClass;
               }
             }
-            if (startValue && endValue) {
+            if (startValue || endValue) {
               if (isSameDay(current, endValue)) {
                 selected = true;
                 isActiveWeek = true;
                 cls += ' ' + selectedEndDateClass;
+              } else if ((startValue === null || startValue === undefined) && current.isBefore(endValue, 'day')) {
+                cls += ' ' + inRangeClass;
+              } else if ((endValue === null || endValue === undefined) && current.isAfter(startValue, 'day')) {
+                cls += ' ' + inRangeClass;
               } else if (current.isAfter(startValue, 'day') && current.isBefore(endValue, 'day')) {
                 cls += ' ' + inRangeClass;
               }
@@ -281,7 +282,8 @@ var DateTBody = (0, _createReactClass2['default'])({
             onClick: disabled ? undefined : props.onSelect.bind(null, current),
             onMouseEnter: disabled ? undefined : props.onDayHover && props.onDayHover.bind(null, current) || undefined,
             role: 'gridcell',
-            title: (0, _util.getTitleString)(current), className: cls
+            title: (0, _util.getTitleString)(current),
+            className: cls
           },
           dateHtml
         ));
@@ -305,8 +307,24 @@ var DateTBody = (0, _createReactClass2['default'])({
       { className: prefixCls + '-tbody' },
       tableHtml
     );
-  }
-});
+  };
 
+  return DateTBody;
+}(_react2['default'].Component);
+
+DateTBody.propTypes = {
+  contentRender: _propTypes2['default'].func,
+  dateRender: _propTypes2['default'].func,
+  disabledDate: _propTypes2['default'].func,
+  prefixCls: _propTypes2['default'].string,
+  selectedValue: _propTypes2['default'].oneOfType([_propTypes2['default'].object, _propTypes2['default'].arrayOf(_propTypes2['default'].object)]),
+  value: _propTypes2['default'].oneOfType([_propTypes2['default'].object, _propTypes2['default'].arrayOf(_propTypes2['default'].object)]),
+  hoverValue: _propTypes2['default'].any,
+  showWeekNumber: _propTypes2['default'].bool,
+  multiple: _propTypes2['default'].bool
+};
+DateTBody.defaultProps = {
+  hoverValue: []
+};
 exports['default'] = DateTBody;
 module.exports = exports['default'];
