@@ -69,7 +69,10 @@ var DecadePanel = function (_React$Component) {
 
     var props = this.props;
     var displayedValue = props.displayedValue;
-    var locale = props.locale;
+    var _props = this.props,
+        locale = _props.locale,
+        renderFooter = _props.renderFooter;
+
     var currentYear = displayedValue.year();
     var startYear = parseInt(currentYear / 100, 10) * 100;
     var preYear = startYear - 10;
@@ -90,6 +93,8 @@ var DecadePanel = function (_React$Component) {
         index++;
       }
     }
+
+    var footer = renderFooter && renderFooter('decade');
 
     var decadesEls = decades.map(function (row, decadeIndex) {
       var tds = row.map(function (decadeData) {
@@ -171,6 +176,11 @@ var DecadePanel = function (_React$Component) {
             decadesEls
           )
         )
+      ),
+      footer && _react2['default'].createElement(
+        'div',
+        { className: prefixCls + '-footer' },
+        footer
       )
     );
   };
@@ -185,7 +195,8 @@ DecadePanel.propTypes = {
   locale: _propTypes2['default'].object,
   value: _propTypes2['default'].object,
   defaultValue: _propTypes2['default'].object,
-  rootPrefixCls: _propTypes2['default'].string
+  rootPrefixCls: _propTypes2['default'].string,
+  renderFooter: _propTypes2['default'].func
 };
 
 DecadePanel.defaultProps = {
