@@ -10,7 +10,7 @@ exports['default'] = {
   month: 'Mois',
   year: 'Année',
   timeSelect: 'Sélectionner l\'heure',
-  dateSelect: 'Sélectionner l\'heure',
+  dateSelect: 'Sélectionner la date',
   monthSelect: 'Choisissez un mois',
   yearSelect: 'Choisissez une année',
   decadeSelect: 'Choisissez une décennie',
