@@ -6,13 +6,21 @@ var _extends2 = require('babel-runtime/helpers/extends');
 
 var _extends3 = _interopRequireDefault(_extends2);
 
+var _classCallCheck2 = require('babel-runtime/helpers/classCallCheck');
+
+var _classCallCheck3 = _interopRequireDefault(_classCallCheck2);
+
+var _possibleConstructorReturn2 = require('babel-runtime/helpers/possibleConstructorReturn');
+
+var _possibleConstructorReturn3 = _interopRequireDefault(_possibleConstructorReturn2);
+
+var _inherits2 = require('babel-runtime/helpers/inherits');
+
+var _inherits3 = _interopRequireDefault(_inherits2);
+
 var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
-
-var _createReactClass = require('create-react-class');
-
-var _createReactClass2 = _interopRequireDefault(_createReactClass);
 
 var _propTypes = require('prop-types');
 
@@ -34,29 +42,15 @@ var _index = require('../util/index');
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
 
-var CalendarPart = (0, _createReactClass2['default'])({
-  displayName: 'CalendarPart',
+var CalendarPart = function (_React$Component) {
+  (0, _inherits3['default'])(CalendarPart, _React$Component);
 
-  propTypes: {
-    prefixCls: _propTypes2['default'].string,
-    value: _propTypes2['default'].any,
-    hoverValue: _propTypes2['default'].any,
-    selectedValue: _propTypes2['default'].any,
-    direction: _propTypes2['default'].any,
-    locale: _propTypes2['default'].any,
-    showDateInput: _propTypes2['default'].bool,
-    showTimePicker: _propTypes2['default'].bool,
-    format: _propTypes2['default'].any,
-    placeholder: _propTypes2['default'].any,
-    disabledDate: _propTypes2['default'].any,
-    timePicker: _propTypes2['default'].any,
-    disabledTime: _propTypes2['default'].any,
-    onInputSelect: _propTypes2['default'].func,
-    timePickerDisabledTime: _propTypes2['default'].object,
-    enableNext: _propTypes2['default'].any,
-    enablePrev: _propTypes2['default'].any
-  },
-  render: function render() {
+  function CalendarPart() {
+    (0, _classCallCheck3['default'])(this, CalendarPart);
+    return (0, _possibleConstructorReturn3['default'])(this, _React$Component.apply(this, arguments));
+  }
+
+  CalendarPart.prototype.render = function render() {
     var props = this.props;
     var prefixCls = props.prefixCls,
         value = props.value,
@@ -72,9 +66,13 @@ var CalendarPart = (0, _createReactClass2['default'])({
         disabledTime = props.disabledTime,
         timePickerDisabledTime = props.timePickerDisabledTime,
         showTimePicker = props.showTimePicker,
+        onInputChange = props.onInputChange,
         onInputSelect = props.onInputSelect,
         enablePrev = props.enablePrev,
-        enableNext = props.enableNext;
+        enableNext = props.enableNext,
+        clearIcon = props.clearIcon,
+        showClear = props.showClear,
+        inputMode = props.inputMode;
 
     var shouldShowTimePicker = showTimePicker && timePicker;
     var disabledTimeConfig = shouldShowTimePicker && disabledTime ? (0, _index.getTimeConfig)(selectedValue, disabledTime) : null;
@@ -92,7 +90,7 @@ var CalendarPart = (0, _createReactClass2['default'])({
       showMinute: true,
       showSecond: true
     }, timePicker.props, disabledTimeConfig, timePickerDisabledTime, {
-      onChange: onInputSelect,
+      onChange: onInputChange,
       defaultOpenValue: value,
       value: selectedValue[index]
     }));
@@ -106,14 +104,20 @@ var CalendarPart = (0, _createReactClass2['default'])({
       placeholder: placeholder,
       disabledTime: disabledTime,
       value: value,
-      showClear: false,
+      showClear: showClear || false,
       selectedValue: selectedValue[index],
-      onChange: onInputSelect
+      onChange: onInputChange,
+      onSelect: onInputSelect,
+      onClear: function onClear() {},
+      clearIcon: clearIcon,
+      inputMode: inputMode
     });
 
     return _react2['default'].createElement(
       'div',
-      { className: rangeClassName + '-part ' + rangeClassName + '-' + direction },
+      {
+        className: rangeClassName + '-part ' + rangeClassName + '-' + direction
+      },
       dateInputElement,
       _react2['default'].createElement(
         'div',
@@ -151,8 +155,33 @@ var CalendarPart = (0, _createReactClass2['default'])({
         )
       )
     );
-  }
-});
+  };
 
+  return CalendarPart;
+}(_react2['default'].Component);
+
+CalendarPart.propTypes = {
+  prefixCls: _propTypes2['default'].string,
+  value: _propTypes2['default'].any,
+  hoverValue: _propTypes2['default'].any,
+  selectedValue: _propTypes2['default'].any,
+  direction: _propTypes2['default'].any,
+  locale: _propTypes2['default'].any,
+  showDateInput: _propTypes2['default'].bool,
+  showTimePicker: _propTypes2['default'].bool,
+  format: _propTypes2['default'].any,
+  placeholder: _propTypes2['default'].any,
+  disabledDate: _propTypes2['default'].any,
+  timePicker: _propTypes2['default'].any,
+  disabledTime: _propTypes2['default'].any,
+  onInputChange: _propTypes2['default'].func,
+  onInputSelect: _propTypes2['default'].func,
+  timePickerDisabledTime: _propTypes2['default'].object,
+  enableNext: _propTypes2['default'].any,
+  enablePrev: _propTypes2['default'].any,
+  clearIcon: _propTypes2['default'].node,
+  dateRender: _propTypes2['default'].func,
+  inputMode: _propTypes2['default'].string
+};
 exports['default'] = CalendarPart;
 module.exports = exports['default'];
