@@ -87,8 +87,10 @@ var YearPanel = function (_React$Component) {
     var _this2 = this;
 
     var props = this.props;
+    var locale = props.locale,
+        renderFooter = props.renderFooter;
+
     var displayedValue = props.displayedValue;
-    var locale = props.locale;
     var years = this.years();
     var currentYear = displayedValue.year();
     var startYear = parseInt(currentYear / 10, 10) * 10;
@@ -132,6 +134,8 @@ var YearPanel = function (_React$Component) {
         tds
       );
     });
+
+    var footer = renderFooter && renderFooter('year');
 
     return _react2['default'].createElement(
       'div',
@@ -188,6 +192,11 @@ var YearPanel = function (_React$Component) {
               yeasEls
             )
           )
+        ),
+        footer && _react2['default'].createElement(
+          'div',
+          { className: prefixCls + '-footer' },
+          footer
         )
       )
     );
@@ -203,6 +212,7 @@ YearPanel.propTypes = {
   rootPrefixCls: _propTypes2['default'].string,
   value: _propTypes2['default'].object,
   defaultValue: _propTypes2['default'].object,
+  renderFooter: _propTypes2['default'].func,
   displayedValue: _propTypes2['default'].object
 };
 
