@@ -1,16 +1,18 @@
 import React from 'react';
-import createReactClass from 'create-react-class';
 import PropTypes from 'prop-types';
 import moment from 'moment';
 import classnames from 'classnames';
+import { polyfill } from 'react-lifecycles-compat';
+import KeyCode from 'rc-util/lib/KeyCode';
 import CalendarPart from './range-calendar/CalendarPart';
 import TodayButton from './calendar/TodayButton';
 import OkButton from './calendar/OkButton';
 import TimePickerButton from './calendar/TimePickerButton';
-import CommonMixin from './mixin/CommonMixin';
-import { syncTime, getTodayTime, isAllowedDate } from './util/';
+import { commonMixinWrapper, propType, defaultProp } from './mixin/CommonMixin';
+import { syncTime, getTodayTime, isAllowedDate } from './util';
+import { goTime, goStartMonth, goEndMonth, includesTime } from './util/toTime';
 
-function noop() {}
+function noop() { }
 
 function isEmptyArray(arr) {
   return Array.isArray(arr) && (arr.length === 0 || arr.every(i => !i));
@@ -30,17 +32,23 @@ function isArraysEqual(a, b) {
 }
 
 function getValueFromSelectedValue(selectedValue) {
-  const [start, end] = selectedValue;
-  const newEnd = end && end.isSame(start, 'month') ? end.clone().add(1, 'month') : end;
-  return [start, newEnd];
+  let [start, end] = selectedValue;
+  if (end && (start === undefined || start === null)) {
+    start = end.clone().subtract(1, 'month');
+  }
+
+  if (start && (end === undefined || end === null)) {
+    end = start.clone().add(1, 'month');
+  }
+  return [start, end];
 }
 
 function normalizeAnchor(props, init) {
   const selectedValue = props.selectedValue || init && props.defaultSelectedValue;
   const value = props.value || init && props.defaultValue;
   const normalizedValue = value ?
-          getValueFromSelectedValue(value) :
-          getValueFromSelectedValue(selectedValue);
+    getValueFromSelectedValue(value) :
+    getValueFromSelectedValue(selectedValue);
   return !isEmptyArray(normalizedValue) ?
     normalizedValue : init && [moment(), moment().add(1, 'months')];
 }
@@ -55,7 +63,7 @@ function generateOptions(length, extraOptionGen) {
   return arr;
 }
 
-function onInputSelect(direction, value) {
+function onInputSelect(direction, value, cause) {
   if (!value) {
     return;
   }
@@ -67,17 +75,19 @@ function onInputSelect(direction, value) {
     selectedValue[1 - index] = this.state.showTimePicker ? selectedValue[index] : undefined;
   }
   this.props.onInputSelect(selectedValue);
-  this.fireSelectValueChange(selectedValue);
+  this.fireSelectValueChange(selectedValue, null, cause || { source: 'dateInput' });
 }
 
-const RangeCalendar = createReactClass({
-  propTypes: {
+class RangeCalendar extends React.Component {
+  static propTypes = {
+    ...propType,
     prefixCls: PropTypes.string,
     dateInputPlaceholder: PropTypes.any,
+    seperator: PropTypes.string,
     defaultValue: PropTypes.any,
     value: PropTypes.any,
     hoverValue: PropTypes.any,
-    mode: PropTypes.arrayOf(PropTypes.oneOf(['date', 'month', 'year', 'decade'])),
+    mode: PropTypes.arrayOf(PropTypes.oneOf(['time', 'date', 'month', 'year', 'decade'])),
     showDateInput: PropTypes.bool,
     timePicker: PropTypes.any,
     showOk: PropTypes.bool,
@@ -97,31 +107,31 @@ const RangeCalendar = createReactClass({
     type: PropTypes.any,
     disabledDate: PropTypes.func,
     disabledTime: PropTypes.func,
+    clearIcon: PropTypes.node,
+    onKeyDown: PropTypes.func,
     highlightToday: PropTypes.bool,
-  },
+  }
 
-  mixins: [CommonMixin],
+  static defaultProps = {
+    ...defaultProp,
+    type: 'both',
+    seperator: '~',
+    defaultSelectedValue: [],
+    onValueChange: noop,
+    onHoverChange: noop,
+    onPanelChange: noop,
+    disabledTime: noop,
+    onInputSelect: noop,
+    showToday: true,
+    showDateInput: true,
+    highlightToday: true,
+  }
 
-  getDefaultProps() {
-    return {
-      type: 'both',
-      defaultSelectedValue: [],
-      onValueChange: noop,
-      onHoverChange: noop,
-      onPanelChange: noop,
-      disabledTime: noop,
-      onInputSelect: noop,
-      showToday: true,
-      showDateInput: true,
-      highlightToday: true,
-    };
-  },
-
-  getInitialState() {
-    const props = this.props;
+  constructor(props) {
+    super(props);
     const selectedValue = props.selectedValue || props.defaultSelectedValue;
     const value = normalizeAnchor(props, 1);
-    return {
+    this.state = {
       selectedValue,
       prevSelectedValue: selectedValue,
       firstSelectedValue: null,
@@ -129,42 +139,23 @@ const RangeCalendar = createReactClass({
       value,
       showTimePicker: false,
       mode: props.mode || ['date', 'date'],
+      panelTriggerSource: '', // Trigger by which picker panel: 'start' & 'end'
     };
-  },
+  }
 
-  componentWillReceiveProps(nextProps) {
-    const { state } = this;
-    const newState = {};
-    if ('value' in nextProps) {
-      newState.value = normalizeAnchor(nextProps, 0);
-      this.setState(newState);
-    }
-    if ('hoverValue' in nextProps && !isArraysEqual(state.hoverValue, nextProps.hoverValue)) {
-      this.setState({ hoverValue: nextProps.hoverValue });
-    }
-    if ('selectedValue' in nextProps) {
-      newState.selectedValue = nextProps.selectedValue;
-      newState.prevSelectedValue = nextProps.selectedValue;
-      this.setState(newState);
-    }
-    if ('mode' in nextProps && !isArraysEqual(state.mode, nextProps.mode)) {
-      this.setState({ mode: nextProps.mode });
-    }
-  },
-
-  onDatePanelEnter() {
+  onDatePanelEnter = () => {
     if (this.hasSelectedValue()) {
       this.fireHoverValueChange(this.state.selectedValue.concat());
     }
-  },
+  }
 
-  onDatePanelLeave() {
+  onDatePanelLeave = () => {
     if (this.hasSelectedValue()) {
       this.fireHoverValueChange([]);
     }
-  },
+  }
 
-  onSelect(value) {
+  onSelect = (value) => {
     const { type } = this.props;
     const { selectedValue, prevSelectedValue, firstSelectedValue } = this.state;
     let nextSelectedValue;
@@ -197,11 +188,135 @@ const RangeCalendar = createReactClass({
     }
 
     this.fireSelectValueChange(nextSelectedValue);
-  },
+  }
 
-  onDayHover(value) {
+  onKeyDown = (event) => {
+    if (event.target.nodeName.toLowerCase() === 'input') {
+      return;
+    }
+
+    const { keyCode } = event;
+    const ctrlKey = event.ctrlKey || event.metaKey;
+
+    const {
+      selectedValue, hoverValue, firstSelectedValue,
+      value, // Value is used for `CalendarPart` current page
+    } = this.state;
+    const { onKeyDown, disabledDate } = this.props;
+
+    // Update last time of the picker
+    const updateHoverPoint = (func) => {
+      // Change hover to make focus in UI
+      let currentHoverTime;
+      let nextHoverTime;
+      let nextHoverValue;
+
+      if (!firstSelectedValue) {
+        currentHoverTime = hoverValue[0] || selectedValue[0] || value[0] || moment();
+        nextHoverTime = func(currentHoverTime);
+        nextHoverValue = [nextHoverTime];
+        this.fireHoverValueChange(nextHoverValue);
+      } else {
+        if (hoverValue.length === 1) {
+          currentHoverTime = hoverValue[0].clone();
+          nextHoverTime = func(currentHoverTime);
+          nextHoverValue = this.onDayHover(nextHoverTime);
+        } else {
+          currentHoverTime = hoverValue[0].isSame(firstSelectedValue, 'day') ?
+            hoverValue[1] : hoverValue[0];
+          nextHoverTime = func(currentHoverTime);
+          nextHoverValue = this.onDayHover(nextHoverTime);
+        }
+      }
+
+      // Find origin hover time on value index
+      if (nextHoverValue.length >= 2) {
+        const miss = nextHoverValue.some(ht => !includesTime(value, ht, 'month'));
+        if (miss) {
+          const newValue = nextHoverValue.slice()
+            .sort((t1, t2) => t1.valueOf() - t2.valueOf());
+          if (newValue[0].isSame(newValue[1], 'month')) {
+            newValue[1] = newValue[0].clone().add(1, 'month');
+          }
+          this.fireValueChange(newValue);
+        }
+      } else if (nextHoverValue.length === 1) {
+        // If only one value, let's keep the origin panel
+        let oriValueIndex = value.findIndex(time => time.isSame(currentHoverTime, 'month'));
+        if (oriValueIndex === -1) oriValueIndex = 0;
+
+        if (value.every(time => !time.isSame(nextHoverTime, 'month'))) {
+          const newValue = value.slice();
+          newValue[oriValueIndex] = nextHoverTime.clone();
+          this.fireValueChange(newValue);
+        }
+      }
+
+      event.preventDefault();
+
+      return nextHoverTime;
+    };
+
+    switch (keyCode) {
+      case KeyCode.DOWN:
+        updateHoverPoint((time) => goTime(time, 1, 'weeks'));
+        return;
+      case KeyCode.UP:
+        updateHoverPoint((time) => goTime(time, -1, 'weeks'));
+        return;
+      case KeyCode.LEFT:
+        if (ctrlKey) {
+          updateHoverPoint((time) => goTime(time, -1, 'years'));
+        } else {
+          updateHoverPoint((time) => goTime(time, -1, 'days'));
+        }
+        return;
+      case KeyCode.RIGHT:
+        if (ctrlKey) {
+          updateHoverPoint((time) => goTime(time, 1, 'years'));
+        } else {
+          updateHoverPoint((time) => goTime(time, 1, 'days'));
+        }
+        return;
+      case KeyCode.HOME:
+        updateHoverPoint((time) => goStartMonth(time));
+        return;
+      case KeyCode.END:
+        updateHoverPoint((time) => goEndMonth(time));
+        return;
+      case KeyCode.PAGE_DOWN:
+        updateHoverPoint((time) => goTime(time, 1, 'month'));
+        return;
+      case KeyCode.PAGE_UP:
+        updateHoverPoint((time) => goTime(time, -1, 'month'));
+        return;
+      case KeyCode.ENTER: {
+        let lastValue;
+        if (hoverValue.length === 0) {
+          lastValue = updateHoverPoint(time => time);
+        } else if (hoverValue.length === 1) {
+          lastValue = hoverValue[0];
+        } else {
+          lastValue = hoverValue[0].isSame(firstSelectedValue, 'day') ?
+            hoverValue[1] : hoverValue[0];
+        }
+        if (lastValue && (!disabledDate || !disabledDate(lastValue))) {
+          this.onSelect(lastValue);
+        }
+        event.preventDefault();
+        return;
+      }
+      default:
+        if (onKeyDown) {
+          onKeyDown(event);
+        }
+    }
+  }
+
+  onDayHover = (value) => {
     let hoverValue = [];
     const { selectedValue, firstSelectedValue } = this.state;
+
     const { type } = this.props;
     if (type === 'start' && selectedValue[1]) {
       hoverValue = this.compare(value, selectedValue[1]) < 0 ?
@@ -211,112 +326,174 @@ const RangeCalendar = createReactClass({
         [selectedValue[0], value] : [];
     } else {
       if (!firstSelectedValue) {
-        return;
+        if (this.state.hoverValue.length) {
+          this.setState({ hoverValue: [] });
+        }
+        return hoverValue;
       }
       hoverValue = this.compare(value, firstSelectedValue) < 0 ?
         [value, firstSelectedValue] : [firstSelectedValue, value];
     }
     this.fireHoverValueChange(hoverValue);
-  },
 
-  onToday() {
+    return hoverValue;
+  }
+
+  onToday = () => {
     const startValue = getTodayTime(this.state.value[0]);
     const endValue = startValue.clone().add(1, 'months');
     this.setState({ value: [startValue, endValue] });
-  },
+  }
 
-  onOpenTimePicker() {
+  onOpenTimePicker = () => {
     this.setState({
       showTimePicker: true,
     });
-  },
-  onCloseTimePicker() {
+  }
+
+  onCloseTimePicker = () => {
     this.setState({
       showTimePicker: false,
     });
-  },
+  }
 
-  onOk() {
+  onOk = () => {
     const { selectedValue } = this.state;
     if (this.isAllowedDateAndTime(selectedValue)) {
       this.props.onOk(this.state.selectedValue);
     }
-  },
+  }
 
-  onStartInputSelect(...oargs) {
+  onStartInputChange = (...oargs) => {
     const args = ['left'].concat(oargs);
     return onInputSelect.apply(this, args);
-  },
+  }
 
-  onEndInputSelect(...oargs) {
+  onEndInputChange = (...oargs) => {
     const args = ['right'].concat(oargs);
     return onInputSelect.apply(this, args);
-  },
+  }
 
-  onStartValueChange(leftValue) {
+  onStartInputSelect = (value) => {
+    const args = ['left', value, { source: 'dateInputSelect' }];
+    return onInputSelect.apply(this, args);
+  }
+
+  onEndInputSelect = (value) => {
+    const args = ['right', value, { source: 'dateInputSelect' }];
+    return onInputSelect.apply(this, args);
+  }
+
+  onStartValueChange = (leftValue) => {
     const value = [...this.state.value];
     value[0] = leftValue;
     return this.fireValueChange(value);
-  },
+  }
 
-  onEndValueChange(rightValue) {
+  onEndValueChange = (rightValue) => {
     const value = [...this.state.value];
     value[1] = rightValue;
     return this.fireValueChange(value);
-  },
+  }
 
-  onStartPanelChange(value, mode) {
+  onStartPanelChange = (value, mode) => {
     const { props, state } = this;
     const newMode = [mode, state.mode[1]];
+    const newState = {
+      panelTriggerSource: 'start',
+    };
     if (!('mode' in props)) {
-      this.setState({
-        mode: newMode,
-      });
+      newState.mode = newMode;
     }
+    this.setState(newState);
     const newValue = [value || state.value[0], state.value[1]];
     props.onPanelChange(newValue, newMode);
-  },
+  }
 
-  onEndPanelChange(value, mode) {
+  onEndPanelChange = (value, mode) => {
     const { props, state } = this;
     const newMode = [state.mode[0], mode];
+    const newState = {
+      panelTriggerSource: 'end',
+    };
     if (!('mode' in props)) {
-      this.setState({
-        mode: newMode,
-      });
+      newState.mode = newMode;
     }
+    this.setState(newState);
     const newValue = [state.value[0], value || state.value[1]];
     props.onPanelChange(newValue, newMode);
-  },
+  }
 
-  getStartValue() {
-    let value = this.state.value[0];
-    const selectedValue = this.state.selectedValue;
+  static getDerivedStateFromProps(nextProps, state) {
+    const newState = {};
+    if ('value' in nextProps) {
+      newState.value = normalizeAnchor(nextProps, 0);
+    }
+    if ('hoverValue' in nextProps && !isArraysEqual(state.hoverValue, nextProps.hoverValue)) {
+      newState.hoverValue = nextProps.hoverValue;
+    }
+    if ('selectedValue' in nextProps) {
+      newState.selectedValue = nextProps.selectedValue;
+      newState.prevSelectedValue = nextProps.selectedValue;
+    }
+    if ('mode' in nextProps && !isArraysEqual(state.mode, nextProps.mode)) {
+      newState.mode = nextProps.mode;
+    }
+    return newState;
+  }
+
+  getStartValue = () => {
+    const { selectedValue, showTimePicker, value, mode, panelTriggerSource } = this.state;
+    let startValue = value[0];
     // keep selectedTime when select date
     if (selectedValue[0] && this.props.timePicker) {
-      value = value.clone();
-      syncTime(selectedValue[0], value);
+      startValue = startValue.clone();
+      syncTime(selectedValue[0], startValue);
     }
-    if (this.state.showTimePicker && selectedValue[0]) {
-      return selectedValue[0];
+    if (showTimePicker && selectedValue[0]) {
+      startValue = selectedValue[0];
     }
-    return value;
-  },
 
-  getEndValue() {
-    const { value, selectedValue, showTimePicker } = this.state;
-    const endValue = value[1] ? value[1].clone() : value[0].clone().add(1, 'month');
+    // Adjust month if date not align
+    if (
+      panelTriggerSource === 'end' &&
+      mode[0] === 'date' &&
+      mode[1] === 'date' &&
+      startValue.isSame(value[1], 'month')
+    ) {
+      startValue = startValue.clone().subtract(1, 'month');
+    }
+
+    return startValue;
+  }
+
+  getEndValue = () => {
+    const { value, selectedValue, showTimePicker, mode, panelTriggerSource } = this.state;
+    let endValue = value[1] ? value[1].clone() : value[0].clone().add(1, 'month');
     // keep selectedTime when select date
     if (selectedValue[1] && this.props.timePicker) {
       syncTime(selectedValue[1], endValue);
     }
     if (showTimePicker) {
-      return selectedValue[1] ? selectedValue[1] : this.getStartValue();
+      endValue = selectedValue[1] ? selectedValue[1] : this.getStartValue();
     }
+
+    // Adjust month if date not align
+    if (
+      !showTimePicker &&
+      panelTriggerSource !== 'end' &&
+      mode[0] === 'date' &&
+      mode[1] === 'date' &&
+      endValue.isSame(value[0], 'month')
+    ) {
+      endValue = endValue.clone().add(1, 'month');
+    }
+
     return endValue;
-  },
+  }
+
   // get disabled hours for second picker
-  getEndDisableTime() {
+  getEndDisableTime = () => {
     const { selectedValue, value } = this.state;
     const { disabledTime } = this.props;
     const userSettingDisabledTime = disabledTime(selectedValue, 'end') || {};
@@ -352,30 +529,30 @@ const RangeCalendar = createReactClass({
       };
     }
     return userSettingDisabledTime;
-  },
+  }
 
-  isAllowedDateAndTime(selectedValue) {
+  isAllowedDateAndTime = (selectedValue) => {
     return isAllowedDate(selectedValue[0], this.props.disabledDate, this.disabledStartTime) &&
-    isAllowedDate(selectedValue[1], this.props.disabledDate, this.disabledEndTime);
-  },
+      isAllowedDate(selectedValue[1], this.props.disabledDate, this.disabledEndTime);
+  }
 
-  isMonthYearPanelShow(mode) {
+  isMonthYearPanelShow = (mode) => {
     return ['month', 'year', 'decade'].indexOf(mode) > -1;
-  },
+  }
 
-  hasSelectedValue() {
+  hasSelectedValue = () => {
     const { selectedValue } = this.state;
     return !!selectedValue[1] && !!selectedValue[0];
-  },
+  }
 
-  compare(v1, v2) {
+  compare = (v1, v2) => {
     if (this.props.timePicker) {
       return v1.diff(v2);
     }
     return v1.diff(v2, 'days');
-  },
+  }
 
-  fireSelectValueChange(selectedValue, direct) {
+  fireSelectValueChange = (selectedValue, direct, cause) => {
     const { timePicker } = this.props;
     const { prevSelectedValue } = this.state;
     if (timePicker && timePicker.props.defaultValue) {
@@ -415,11 +592,11 @@ const RangeCalendar = createReactClass({
         firstSelectedValue: null,
       });
       this.fireHoverValueChange([]);
-      this.props.onSelect(selectedValue);
+      this.props.onSelect(selectedValue, cause);
     }
-  },
+  }
 
-  fireValueChange(value) {
+  fireValueChange = (value) => {
     const props = this.props;
     if (!('value' in props)) {
       this.setState({
@@ -427,45 +604,45 @@ const RangeCalendar = createReactClass({
       });
     }
     props.onValueChange(value);
-  },
+  }
 
-  fireHoverValueChange(hoverValue) {
+  fireHoverValueChange = (hoverValue) => {
     const props = this.props;
     if (!('hoverValue' in props)) {
       this.setState({ hoverValue });
     }
     props.onHoverChange(hoverValue);
-  },
+  }
 
-  clear() {
+  clear = () => {
     this.fireSelectValueChange([], true);
     this.props.onClear();
-  },
+  }
 
-  disabledStartTime(time) {
+  disabledStartTime = (time) => {
     return this.props.disabledTime(time, 'start');
-  },
+  }
 
-  disabledEndTime(time) {
+  disabledEndTime = (time) => {
     return this.props.disabledTime(time, 'end');
-  },
+  }
 
-  disabledStartMonth(month) {
+  disabledStartMonth = (month) => {
     const { value } = this.state;
-    return month.isSameOrAfter(value[1], 'month');
-  },
+    return month.isAfter(value[1], 'month');
+  }
 
-  disabledEndMonth(month) {
+  disabledEndMonth = (month) => {
     const { value } = this.state;
-    return month.isSameOrBefore(value[0], 'month');
-  },
+    return month.isBefore(value[0], 'month');
+  }
 
   render() {
     const { props, state } = this;
     const {
-      prefixCls, dateInputPlaceholder,
+      prefixCls, dateInputPlaceholder, seperator,
       timePicker, showOk, locale, showClear,
-      showToday, type,
+      showToday, type, clearIcon,
     } = props;
     const {
       hoverValue,
@@ -486,7 +663,7 @@ const RangeCalendar = createReactClass({
       selectedValue: state.selectedValue,
       onSelect: this.onSelect,
       onDayHover: type === 'start' && selectedValue[1] ||
-      type === 'end' && selectedValue[0] || !!hoverValue.length ?
+        type === 'end' && selectedValue[0] || !!hoverValue.length ?
         this.onDayHover : undefined,
     };
 
@@ -513,27 +690,32 @@ const RangeCalendar = createReactClass({
     const thisMonth = todayTime.month();
     const thisYear = todayTime.year();
     const isTodayInView =
-            startValue.year() === thisYear && startValue.month() === thisMonth ||
-            endValue.year() === thisYear && endValue.month() === thisMonth;
+      startValue.year() === thisYear && startValue.month() === thisMonth ||
+      endValue.year() === thisYear && endValue.month() === thisMonth;
     const nextMonthOfStart = startValue.clone().add(1, 'months');
     const isClosestMonths = nextMonthOfStart.year() === endValue.year() &&
-            nextMonthOfStart.month() === endValue.month();
+      nextMonthOfStart.month() === endValue.month();
+
+    const extraFooter = props.renderFooter();
+
     return (
       <div
         ref={this.saveRoot}
         className={classes}
         style={props.style}
         tabIndex="0"
+        onKeyDown={this.onKeyDown}
       >
         {props.renderSidebar()}
         <div className={`${prefixCls}-panel`}>
           {showClear && selectedValue[0] && selectedValue[1] ?
             <a
-              className={`${prefixCls}-clear-btn`}
               role="button"
               title={locale.clear}
               onClick={this.clear}
-            /> : null}
+            >
+              {clearIcon || <span className={`${prefixCls}-clear-btn`} />}
+            </a> : null}
           <div
             className={`${prefixCls}-date-panel`}
             onMouseLeave={type !== 'both' ? this.onDatePanelLeave : undefined}
@@ -550,16 +732,18 @@ const RangeCalendar = createReactClass({
               value={startValue}
               mode={mode[0]}
               placeholder={placeholder1}
+              onInputChange={this.onStartInputChange}
               onInputSelect={this.onStartInputSelect}
               onValueChange={this.onStartValueChange}
               onPanelChange={this.onStartPanelChange}
               showDateInput={this.props.showDateInput}
               timePicker={timePicker}
-              showTimePicker={showTimePicker}
+              showTimePicker={showTimePicker || mode[0] === 'time'}
               enablePrev
               enableNext={!isClosestMonths || this.isMonthYearPanelShow(mode[1])}
+              clearIcon={clearIcon}
             />
-            <span className={`${prefixCls}-range-middle`}>~</span>
+            <span className={`${prefixCls}-range-middle`}>{seperator}</span>
             <CalendarPart
               {...props}
               {...newProps}
@@ -570,22 +754,24 @@ const RangeCalendar = createReactClass({
               placeholder={placeholder2}
               value={endValue}
               mode={mode[1]}
+              onInputChange={this.onEndInputChange}
               onInputSelect={this.onEndInputSelect}
               onValueChange={this.onEndValueChange}
               onPanelChange={this.onEndPanelChange}
               showDateInput={this.props.showDateInput}
               timePicker={timePicker}
-              showTimePicker={showTimePicker}
+              showTimePicker={showTimePicker || mode[1] === 'time'}
               disabledTime={this.disabledEndTime}
               disabledMonth={this.disabledEndMonth}
               enablePrev={!isClosestMonths || this.isMonthYearPanelShow(mode[0])}
               enableNext
+              clearIcon={clearIcon}
             />
           </div>
           <div className={cls}>
-            {props.renderFooter()}
-            {showToday || props.timePicker || showOkButton ? (
+            {(showToday || props.timePicker || showOkButton || extraFooter) ? (
               <div className={`${prefixCls}-footer-btn`}>
+                {extraFooter}
                 {showToday ? (
                   <TodayButton
                     {...props}
@@ -598,7 +784,7 @@ const RangeCalendar = createReactClass({
                 {props.timePicker ?
                   <TimePickerButton
                     {...props}
-                    showTimePicker={showTimePicker}
+                    showTimePicker={showTimePicker || (mode[0] === 'time' && mode[1] === 'time')}
                     onOpenTimePicker={this.onOpenTimePicker}
                     onCloseTimePicker={this.onCloseTimePicker}
                     timePickerDisabled={!this.hasSelectedValue() || hoverValue.length}
@@ -617,7 +803,9 @@ const RangeCalendar = createReactClass({
         </div>
       </div>
     );
-  },
-});
+  }
+}
 
-export default RangeCalendar;
+polyfill(RangeCalendar);
+
+export default commonMixinWrapper(RangeCalendar);
