@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import createReactClass from 'create-react-class';
 import PropTypes from 'prop-types';
 import toFragment from 'rc-util/lib/Children/mapSelf';
 import cx from 'classnames';
@@ -8,8 +7,8 @@ import TodayButton from '../calendar/TodayButton';
 import OkButton from '../calendar/OkButton';
 import TimePickerButton from '../calendar/TimePickerButton';
 
-const CalendarFooter = createReactClass({
-  propTypes: {
+export default class CalendarFooter extends React.Component {
+  static propTypes = {
     prefixCls: PropTypes.string,
     showDateInput: PropTypes.bool,
     disabledTime: PropTypes.any,
@@ -20,23 +19,24 @@ const CalendarFooter = createReactClass({
     value: PropTypes.oneOfType([PropTypes.object, PropTypes.arrayOf(PropTypes.object)]),
     displayedValue: PropTypes.object,
     renderFooter: PropTypes.func,
+    mode: PropTypes.string,
     defaultValue: PropTypes.oneOfType([PropTypes.object, PropTypes.arrayOf(PropTypes.object)]),
     multiple: PropTypes.bool,
-  },
+  }
 
   onSelect(value) {
     this.props.onSelect(value);
-  },
+  }
 
   getRootDOMNode() {
     return ReactDOM.findDOMNode(this);
-  },
+  }
 
   render() {
     const props = this.props;
-    const { value, prefixCls, showOk, timePicker, renderFooter } = props;
+    const { value, prefixCls, showOk, timePicker, renderFooter, mode } = props;
     let footerEl = null;
-    const extraFooter = renderFooter();
+    const extraFooter = renderFooter && renderFooter(mode);
     if (props.showToday || timePicker || extraFooter) {
       let nowEl;
       if (props.showToday) {
@@ -52,23 +52,23 @@ const CalendarFooter = createReactClass({
       }
 
       let footerBtn;
-      if (nowEl || timePickerBtn || okBtn) {
-        footerBtn = (<span className={`${prefixCls}-footer-btn`}>
-          {toFragment([nowEl, timePickerBtn, okBtn])}
-        </span>);
+      if (nowEl || timePickerBtn || okBtn || extraFooter) {
+        footerBtn = (
+          <span className={`${prefixCls}-footer-btn`}>
+            {extraFooter}
+            {toFragment([nowEl, timePickerBtn, okBtn])}
+          </span>
+        );
       }
-      const cls = cx({
-        [`${prefixCls}-footer`]: true,
+      const cls = cx(`${prefixCls}-footer`, {
         [`${prefixCls}-footer-show-ok`]: okBtn,
       });
       footerEl = (
         <div className={cls}>
-          {extraFooter}
           {footerBtn}
-        </div>);
+        </div>
+      );
     }
     return footerEl;
-  },
-});
-
-export default CalendarFooter;
+  }
+}
