@@ -1,50 +1,162 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import createReactClass from 'create-react-class';
 import PropTypes from 'prop-types';
+import KeyCode from 'rc-util/lib/KeyCode';
+import { polyfill } from 'react-lifecycles-compat';
 import moment from 'moment';
+import { formatDate } from '../util';
 
-const DateInput = createReactClass({
-  propTypes: {
+let cachedSelectionStart;
+let cachedSelectionEnd;
+let dateInputInstance;
+
+class DateInput extends React.Component {
+  static propTypes = {
     prefixCls: PropTypes.string,
     timePicker: PropTypes.object,
     value: PropTypes.oneOfType([PropTypes.object, PropTypes.arrayOf(PropTypes.object)]),
     disabledTime: PropTypes.any,
-    format: PropTypes.string,
+    format: PropTypes.oneOfType([PropTypes.string, PropTypes.arrayOf(PropTypes.string)]),
     locale: PropTypes.object,
     disabledDate: PropTypes.func,
     onChange: PropTypes.func,
-    onClear: PropTypes.func,
+    onClear: PropTypes.func.isRequired,
     placeholder: PropTypes.string,
     onSelect: PropTypes.func,
     selectedValue: PropTypes.oneOfType([PropTypes.object, PropTypes.arrayOf(PropTypes.object)]),
+    clearIcon: PropTypes.node,
+    inputMode: PropTypes.string,
     multiple: PropTypes.bool,
-  },
+  }
 
-  getInitialState() {
-    return {
-      str: this.formatStr(this.props),
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      str: DateInput.formatStr(this.props),
       invalid: false,
+      hasFocus: false,
     };
-  },
-
-  componentWillReceiveProps(nextProps) {
-    this.cachedSelectionStart = this.dateInputInstance.selectionStart;
-    this.cachedSelectionEnd = this.dateInputInstance.selectionEnd;
-    // when popup show, click body will call this, bug!
-    this.setState({
-      str: this.formatStr(nextProps),
-      invalid: false,
-    });
-  },
+  }
 
   componentDidUpdate() {
-    if (!this.state.invalid) {
-      this.dateInputInstance.setSelectionRange(this.cachedSelectionStart, this.cachedSelectionEnd);
+    if (dateInputInstance && this.state.hasFocus && !this.state.invalid &&
+      !(cachedSelectionStart === 0 && cachedSelectionEnd === 0)) {
+      dateInputInstance.setSelectionRange(cachedSelectionStart, cachedSelectionEnd);
     }
-  },
+  }
 
-  formatStr(props) {
+  onClear = () => {
+    this.setState({
+      str: '',
+    });
+    this.props.onClear(null);
+  }
+
+  onInputChange = (event) => {
+    if (this.props.multiple) return;
+
+    const str = event.target.value;
+    const { disabledDate, format, onChange, selectedValue } = this.props;
+
+    // 没有内容，合法并直接退出
+    if (!str) {
+      onChange(null);
+      this.setState({
+        invalid: false,
+        str,
+      });
+      return;
+    }
+
+    // 不合法直接退出
+    const parsed = moment(str, format, true);
+    if (!parsed.isValid()) {
+      this.setState({
+        invalid: true,
+        str,
+      });
+      return;
+    }
+
+    const value = this.props.value.clone();
+    value
+      .year(parsed.year())
+      .month(parsed.month())
+      .date(parsed.date())
+      .hour(parsed.hour())
+      .minute(parsed.minute())
+      .second(parsed.second());
+
+    if (!value || (disabledDate && disabledDate(value))) {
+      this.setState({
+        invalid: true,
+        str,
+      });
+      return;
+    }
+
+    if (selectedValue !== value || (
+      selectedValue && value && !selectedValue.isSame(value)
+    )) {
+      this.setState({
+        invalid: false,
+        str,
+      });
+      onChange(value);
+    }
+  }
+
+  onFocus = () => {
+    this.setState({ hasFocus: true });
+  }
+
+  onBlur = () => {
+    this.setState((prevState, prevProps) => ({
+      hasFocus: false,
+      str: formatDate(prevProps.value, prevProps.format),
+    }));
+  }
+
+  onKeyDown = (event) => {
+    const { keyCode } = event;
+    const { onSelect, value, disabledDate } = this.props;
+    if (keyCode === KeyCode.ENTER && onSelect) {
+      const validateDate = !disabledDate || !disabledDate(value);
+      if (validateDate) {
+        onSelect(value.clone());
+      }
+      event.preventDefault();
+    }
+  };
+
+  static getDerivedStateFromProps(nextProps, state) {
+    let newState = {};
+
+    if (dateInputInstance) {
+      cachedSelectionStart = dateInputInstance.selectionStart;
+      cachedSelectionEnd = dateInputInstance.selectionEnd;
+    }
+    // when popup show, click body will call this, bug!
+    if (!state.hasFocus) {
+      newState = {
+        str: DateInput.formatStr(nextProps),
+        invalid: false,
+      };
+    }
+
+    return newState;
+  }
+
+  static getInstance() {
+    return dateInputInstance;
+  }
+
+  getRootDOMNode = () => {
+    return ReactDOM.findDOMNode(this);
+  }
+
+  static formatStr(props) {
     let str;
 
     const {
@@ -55,109 +167,60 @@ const DateInput = createReactClass({
 
     if (multiple) {
       str = selectedValue && selectedValue.length && selectedValue.map((singleValue) => {
-        return singleValue.format(format) || '';
+        return formatDate(singleValue, format);
       }).join(', ') || '';
     } else {
-      str = selectedValue && selectedValue.format(format) || '';
+      str = formatDate(selectedValue, format);
     }
 
     return str;
-  },
+  }
 
-  onInputChange(event) {
-    if (this.props.multiple) return;
-
-    const str = event.target.value;
-    this.setState({
-      str,
-    });
-    let value;
-    const { disabledDate, format, onChange } = this.props;
-    if (str) {
-      const parsed = moment(str, format, true);
-      if (!parsed.isValid()) {
-        this.setState({
-          invalid: true,
-        });
-        return;
-      }
-      value = this.props.value.clone();
-      value
-        .year(parsed.year())
-        .month(parsed.month())
-        .date(parsed.date())
-        .hour(parsed.hour())
-        .minute(parsed.minute())
-        .second(parsed.second());
-
-      if (value && (!disabledDate || !disabledDate(value))) {
-        const originalValue = this.props.selectedValue;
-        if (originalValue && value) {
-          if (!originalValue.isSame(value)) {
-            onChange(value);
-          }
-        } else if (originalValue !== value) {
-          onChange(value);
-        }
-      } else {
-        this.setState({
-          invalid: true,
-        });
-        return;
-      }
-    } else {
-      onChange(null);
+  focus = () => {
+    if (dateInputInstance) {
+      dateInputInstance.focus();
     }
-    this.setState({
-      invalid: false,
-    });
-  },
+  }
 
-  onClear() {
-    this.setState({
-      str: '',
-    });
-    this.props.onClear(null);
-  },
-
-  getRootDOMNode() {
-    return ReactDOM.findDOMNode(this);
-  },
-
-  focus() {
-    if (this.dateInputInstance) {
-      this.dateInputInstance.focus();
-    }
-  },
-
-  saveDateInput(dateInput) {
-    this.dateInputInstance = dateInput;
-  },
+  saveDateInput = (dateInput) => {
+    dateInputInstance = dateInput;
+  }
 
   render() {
     const props = this.props;
     const { invalid, str } = this.state;
-    const { locale, prefixCls, placeholder } = props;
+    const { locale, prefixCls, placeholder, clearIcon, inputMode } = props;
     const invalidClass = invalid ? `${prefixCls}-input-invalid` : '';
-    return (<div className={`${prefixCls}-input-wrap`}>
-      <div className={`${prefixCls}-date-input-wrap`}>
-        <input
-          ref={this.saveDateInput}
-          className={`${prefixCls}-input ${invalidClass}`}
-          value={str}
-          disabled={props.disabled}
-          placeholder={placeholder}
-          onChange={this.onInputChange}
-        />
+    return (
+      <div className={`${prefixCls}-input-wrap`}>
+        <div className={`${prefixCls}-date-input-wrap`}>
+          <input
+            ref={this.saveDateInput}
+            className={`${prefixCls}-input ${invalidClass}`}
+            value={str}
+            disabled={props.disabled}
+            placeholder={placeholder}
+            onChange={this.onInputChange}
+            onKeyDown={this.onKeyDown}
+            onFocus={this.onFocus}
+            onBlur={this.onBlur}
+            inputMode={inputMode}
+          />
+        </div>
+        {props.showClear ? (
+          <a
+            role="button"
+            title={locale.clear}
+            onClick={this.onClear}
+          >
+            {clearIcon || <span className={`${prefixCls}-clear-btn`} />}
+          </a>
+        ) : null}
       </div>
-      {props.showClear ? <a
-        className={`${prefixCls}-clear-btn`}
-        role="button"
-        title={locale.clear}
-        onClick={this.onClear}
-      /> : null}
-    </div>);
-  },
-});
+    );
+  }
+}
+
+polyfill(DateInput);
 
 export default DateInput;
