@@ -1,5 +1,4 @@
 import React from 'react';
-import createReactClass from 'create-react-class';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 import DateConstants from './DateConstants';
@@ -29,8 +28,8 @@ function getIdFromDate(date) {
   return `rc-calendar-${date.year()}-${date.month()}-${date.date()}`;
 }
 
-const DateTBody = createReactClass({
-  propTypes: {
+export default class DateTBody extends React.Component {
+  static propTypes = {
     contentRender: PropTypes.func,
     dateRender: PropTypes.func,
     disabledDate: PropTypes.func,
@@ -40,13 +39,11 @@ const DateTBody = createReactClass({
     hoverValue: PropTypes.any,
     showWeekNumber: PropTypes.bool,
     multiple: PropTypes.bool,
-  },
+  }
 
-  getDefaultProps() {
-    return {
-      hoverValue: [],
-    };
-  },
+  static defaultProps = {
+    hoverValue: [],
+  }
 
   render() {
     const props = this.props;
@@ -88,6 +85,7 @@ const DateTBody = createReactClass({
     const lastMonth1 = month1.clone();
     lastMonth1.add(0 - lastMonthDiffDay, 'days');
     let passed = 0;
+
     for (iIndex = 0; iIndex < DateConstants.DATE_ROW_COUNT; iIndex++) {
       for (jIndex = 0; jIndex < DateConstants.DATE_COL_COUNT; jIndex++) {
         current = lastMonth1;
@@ -175,11 +173,17 @@ const DateTBody = createReactClass({
                 cls += ` ${selectedStartDateClass}`;
               }
             }
-            if (startValue && endValue) {
+            if (startValue || endValue) {
               if (isSameDay(current, endValue)) {
                 selected = true;
                 isActiveWeek = true;
                 cls += ` ${selectedEndDateClass}`;
+              } else if ((startValue === null || startValue === undefined) &&
+                current.isBefore(endValue, 'day')) {
+                cls += ` ${inRangeClass}`;
+              } else if ((endValue === null || endValue === undefined) &&
+                current.isAfter(startValue, 'day')) {
+                cls += ` ${inRangeClass}`;
               } else if (current.isAfter(startValue, 'day') &&
                 current.isBefore(endValue, 'day')) {
                 cls += ` ${inRangeClass}`;
@@ -255,7 +259,8 @@ const DateTBody = createReactClass({
             onMouseEnter={disabled ?
               undefined : props.onDayHover && props.onDayHover.bind(null, current) || undefined}
             role="gridcell"
-            title={getTitleString(current)} className={cls}
+            title={getTitleString(current)}
+            className={cls}
           >
             {dateHtml}
           </td>);
@@ -281,7 +286,5 @@ const DateTBody = createReactClass({
     return (<tbody className={`${prefixCls}-tbody`}>
       {tableHtml}
     </tbody>);
-  },
-});
-
-export default DateTBody;
+  }
+}
