@@ -3,8 +3,7 @@ import React from 'react';
 import DateTHead from './DateTHead';
 import DateTBody from './DateTBody';
 
-export default
-class DateTable extends React.Component {
+export default class DateTable extends React.Component {
   render() {
     const props = this.props;
     const prefixCls = props.prefixCls;
