@@ -1,16 +1,16 @@
 export default {
   today: 'Oggi',
-  now: 'adesso',
+  now: 'Adesso',
   backToToday: 'Torna ad oggi',
   ok: 'Ok',
-  clear: 'Chiaro',
+  clear: 'Cancella',
   month: 'Mese',
   year: 'Anno',
-  timeSelect: 'Seleziona il tempo',
-  dateSelect: 'Select date',
-  monthSelect: 'Seleziona la data',
-  yearSelect: 'Scegli un anno',
-  decadeSelect: 'Scegli un decennio',
+  timeSelect: 'Seleziona l\'ora',
+  dateSelect: 'Seleziona la data',
+  monthSelect: 'Seleziona il mese',
+  yearSelect: 'Seleziona l\'anno',
+  decadeSelect: 'Seleziona il decennio',
   yearFormat: 'YYYY',
   dateFormat: 'D/M/YYYY',
   dayFormat: 'D',
