@@ -1,13 +1,12 @@
 import React from 'react';
-import createReactClass from 'create-react-class';
 import PropTypes from 'prop-types';
 import CalendarHeader from '../calendar/CalendarHeader';
 import DateTable from '../date/DateTable';
 import DateInput from '../date/DateInput';
 import { getTimeConfig } from '../util/index';
 
-const CalendarPart = createReactClass({
-  propTypes: {
+export default class CalendarPart extends React.Component {
+  static propTypes = {
     prefixCls: PropTypes.string,
     value: PropTypes.any,
     hoverValue: PropTypes.any,
@@ -21,11 +20,16 @@ const CalendarPart = createReactClass({
     disabledDate: PropTypes.any,
     timePicker: PropTypes.any,
     disabledTime: PropTypes.any,
+    onInputChange: PropTypes.func,
     onInputSelect: PropTypes.func,
     timePickerDisabledTime: PropTypes.object,
     enableNext: PropTypes.any,
     enablePrev: PropTypes.any,
-  },
+    clearIcon: PropTypes.node,
+    dateRender: PropTypes.func,
+    inputMode: PropTypes.string,
+  }
+
   render() {
     const props = this.props;
     const {
@@ -38,7 +42,10 @@ const CalendarPart = createReactClass({
       locale, format, placeholder,
       disabledDate, timePicker, disabledTime,
       timePickerDisabledTime, showTimePicker,
-      onInputSelect, enablePrev, enableNext,
+      onInputChange, onInputSelect, enablePrev, enableNext,
+      clearIcon,
+      showClear,
+      inputMode,
     } = props;
     const shouldShowTimePicker = showTimePicker && timePicker;
     const disabledTimeConfig = shouldShowTimePicker && disabledTime ?
@@ -60,7 +67,7 @@ const CalendarPart = createReactClass({
         ...timePicker.props,
         ...disabledTimeConfig,
         ...timePickerDisabledTime,
-        onChange: onInputSelect,
+        onChange: onInputChange,
         defaultOpenValue: value,
         value: selectedValue[index],
       });
@@ -75,13 +82,19 @@ const CalendarPart = createReactClass({
         placeholder={placeholder}
         disabledTime={disabledTime}
         value={value}
-        showClear={false}
+        showClear={showClear || false}
         selectedValue={selectedValue[index]}
-        onChange={onInputSelect}
+        onChange={onInputChange}
+        onSelect={onInputSelect}
+        onClear={() => {}}
+        clearIcon={clearIcon}
+        inputMode={inputMode}
       />;
 
     return (
-      <div className={`${rangeClassName}-part ${rangeClassName}-${direction}`}>
+      <div
+        className={`${rangeClassName}-part ${rangeClassName}-${direction}`}
+      >
         {dateInputElement}
         <div style={{ outline: 'none' }}>
           <CalendarHeader
@@ -113,7 +126,5 @@ const CalendarPart = createReactClass({
           </div>
         </div>
       </div>);
-  },
-});
-
-export default CalendarPart;
+  }
+}
